+++ packages/spotlight/src/spotlight.js
@@ -602,32 +602,28 @@ const Spotlight = (function () {
 		return true;
 	}
 
-	function getContainerId (elem) {
-		const containers = [..._containers.keys()];
-
-		for (let i = containers.length; i-- > 0;) {
-			const id = containers[i];
-			const config = _containers.get(id);
-			if (!config.selectorDisabled && matchSelector(elem, config.selector)) {
-				return id;
-			}
-		}
-	}
-
-	// returns an array of ids for containers that wrap the element, in order of outer-to-inner, with
+	// returns an object of container data for the supplied element, consisting of 2 properties:
+	// `containerIds` - array of ids for containers that wrap the element, in order of outer-to-inner, with
 	// the last array item being the immediate container id of the element.
+	// `activeContainerId` - id of the container that serves as the active container of the element. Usually
+	// this will be the immediate container id of the element, except in nested container cases where this value
+	// will be the nearest parent container that uses a `restrict` config value of `'self-only'`, if available.
 	function getContainerIds (elem) {
 		const containerIds = [..._containers.keys()];
 		const matches = [];
+		let activeContainerId;
 
 		for (let i = 0, containers = containerIds.length; i < containers; ++i) {
 			const id = containerIds[i];
 			const config = _containers.get(id);
 			if (!config.selectorDisabled && matchSelector(elem, config.selector)) {
 				matches.push(id);
+				if (config.restrict === 'self-only') {
+					activeContainerId = id;
+				}
 			}
 		}
-		return matches;
+		return {containerIds: matches, activeContainerId: activeContainerId || last(matches)};
 	}
 
 	function getContainerNavigableElements (containerId) {
@@ -664,7 +660,7 @@ const Spotlight = (function () {
 		}
 	}
 
-	function focusElement (elem, containerIds, fromPointer) {
+	function focusElement (elem, containerIds, activeContainerId, fromPointer) {
 		if (!elem) {
 			return false;
 		}
@@ -681,7 +677,7 @@ const Spotlight = (function () {
 				currentFocusedElement.blur();
 			}
 			elem.focus();
-			focusChanged(elem, containerIds);
+			focusChanged(elem, containerIds, activeContainerId);
 		};
 
 		if (_duringFocusChange) {
@@ -705,18 +701,18 @@ const Spotlight = (function () {
 
 		_duringFocusChange = false;
 
-		focusChanged(elem, containerIds);
+		focusChanged(elem, containerIds, activeContainerId);
 		return true;
 	}
 
-	function focusChanged (elem, containerIds) {
+	function focusChanged (elem, containerIds, activeContainerId) {
 		if (!containerIds || !containerIds.length) {
-			containerIds = getContainerIds(elem);
+			[containerIds, activeContainerId] = getContainerIds(elem);
 		}
 		const containerId = last(containerIds);
 		if (containerId) {
 			setContainerLastFocusedElement(elem, containerIds);
-			_lastContainerId = containerId;
+			_lastContainerId = activeContainerId;
 		}
 	}
 
@@ -731,9 +727,9 @@ const Spotlight = (function () {
 		} else {
 			let next = parseSelector(selector)[0];
 			if (next) {
-				const nextContainerIds = getContainerIds(next);
-				if (isNavigable(next, last(nextContainerIds))) {
-					return focusElement(next, nextContainerIds);
+				const {containerIds, activeContainerId} = getContainerIds(next);
+				if (isNavigable(next, last(containerIds))) {
+					return focusElement(next, containerIds, activeContainerId);
 				}
 			}
 		}
@@ -773,7 +769,8 @@ const Spotlight = (function () {
 			}
 
 			if (next) {
-				return focusElement(next, range);
+				const {containerIds, activeContainerId} = getContainerIds(next);
+				return focusElement(next, containerIds, activeContainerId);
 			}
 		}
 
@@ -793,9 +790,9 @@ const Spotlight = (function () {
 				return focusExtendedSelector(next);
 			}
 
-			const nextContainerIds = getContainerIds(next);
-			if (isNavigable(next, last(nextContainerIds))) {
-				return focusElement(next, nextContainerIds);
+			const {containerIds, activeContainerId} = getContainerIds(next);
+			if (isNavigable(next, last(containerIds))) {
+				return focusElement(next, containerIds, activeContainerId);
 			}
 		}
 		return false;
@@ -813,8 +810,8 @@ const Spotlight = (function () {
 	}
 
 	function focusNext (next, direction, currentContainerId, currentFocusedElement) {
-		const nextContainerIds = getContainerIds(next);
-		const nextContainerId = last(nextContainerIds);
+		const {containerIds, activeContainerId} = getContainerIds(next);
+		const nextContainerId = last(containerIds);
 
 		if (currentContainerId !== nextContainerId) {
 			if (_5WayKeyHold) {
@@ -843,7 +840,7 @@ const Spotlight = (function () {
 			}
 		}
 
-		return focusElement(next, nextContainerIds);
+		return focusElement(next, containerIds, activeContainerId);
 	}
 
 	function spotNextFromPoint (direction, position, containerId) {
@@ -879,7 +876,7 @@ const Spotlight = (function () {
 		return false;
 	}
 
-	function spotNext (direction, currentFocusedElement, currentContainerId) {
+	function spotNext (direction, currentFocusedElement, currentContainerId, currentActiveContainerId) {
 		const extSelector = currentFocusedElement.getAttribute('data-spot-' + direction);
 		if (typeof extSelector === 'string') {
 			if (extSelector === '' || !focusExtendedSelector(extSelector)) {
@@ -890,10 +887,11 @@ const Spotlight = (function () {
 
 		const {allNavigableElements, containerNavigableElements} = getNavigableElements();
 		const config = extend({}, GlobalConfig, _containers.get(currentContainerId));
+		const useActiveContainer = currentContainerId !== currentActiveContainerId;
 		let next;
 
-		if (config.restrict === 'self-only' || config.restrict === 'self-first') {
-			let currentContainerNavigableElements = containerNavigableElements[currentContainerId];
+		if (useActiveContainer || config.restrict === 'self-only' || config.restrict === 'self-first') {
+			const currentContainerNavigableElements = containerNavigableElements[currentContainerId];
 
 			next = navigate(
 				currentFocusedElement,
@@ -902,13 +900,22 @@ const Spotlight = (function () {
 				config
 			);
 
-			if (!next && config.restrict === 'self-first') {
-				next = navigate(
-					currentFocusedElement,
-					direction,
-					exclude(allNavigableElements, currentContainerNavigableElements),
-					config
-				);
+			if (!next) {
+				if (useActiveContainer) {
+					next = navigate(
+						currentFocusedElement,
+						direction,
+						exclude(containerNavigableElements[currentActiveContainerId], currentFocusedElement),
+						extend({}, GlobalConfig, _containers.get(currentActiveContainerId))
+					);
+				} else if (config.restrict === 'self-first') {
+					next = navigate(
+						currentFocusedElement,
+						direction,
+						exclude(allNavigableElements, currentContainerNavigableElements),
+						config
+					);
+				}
 			}
 		} else {
 			next = navigate(
@@ -964,14 +971,14 @@ const Spotlight = (function () {
 			}
 		}
 
-		const currentContainerIds = getContainerIds(currentFocusedElement);
-		const currentContainerId = last(currentContainerIds);
+		const {containerIds, activeContainerId} = getContainerIds(currentFocusedElement);
+		const currentContainerId = last(containerIds);
 		if (!currentContainerId) {
 			return;
 		}
 
-		if (direction && !spotNext(direction, currentFocusedElement, currentContainerId) && currentFocusedElement !== document.activeElement) {
-			focusElement(currentFocusedElement, currentContainerIds);
+		if (direction && !spotNext(direction, currentFocusedElement, currentContainerId, activeContainerId) && currentFocusedElement !== document.activeElement) {
+			focusElement(currentFocusedElement, containerIds, activeContainerId);
 		}
 	}
 
@@ -1036,8 +1043,13 @@ const Spotlight = (function () {
 		const target = getNavigableTarget(evt.target); // account for child controls
 
 		if (target && target !== getCurrent()) { // moving over a focusable element
-			focusElement(target, getContainerIds(target), true);
-			preventDefault(evt);
+			const spotlightModal = _containers.get(_lastContainerId).restrict === 'self-only';
+
+			if (!spotlightModal || (spotlightModal && isNavigable(target, _lastContainerId, true))) {
+				const {containerIds, activeContainerId} = getContainerIds(target);
+				focusElement(target, containerIds, activeContainerId, true);
+				preventDefault(evt);
+			}
 		}
 	}
 
@@ -1074,8 +1086,13 @@ const Spotlight = (function () {
 				// we are moving over a non-focusable element, so we force a blur to occur
 				current.blur();
 			} else if (target && (!current || target !== current)) {
-				// we are moving over a focusable element, so we set focus to the target
-				focusElement(target, getContainerIds(target), true);
+				const spotlightModal = _containers.get(_lastContainerId).restrict === 'self-only';
+
+				if (!spotlightModal || (spotlightModal && isNavigable(target, _lastContainerId, true))) {
+					const {containerIds, activeContainerId} = getContainerIds(target);
+					// we are moving over a focusable element, so we set focus to the target
+					focusElement(target, containerIds, activeContainerId, true);
+				}
 			}
 		}
 	}
@@ -1335,10 +1352,9 @@ const Spotlight = (function () {
 					result = focusExtendedSelector(elem);
 				}
 			} else {
-				const nextContainerIds = getContainerIds(elem);
-				const nextContainerId = last(nextContainerIds);
-				if (isNavigable(elem, nextContainerId)) {
-					result = focusElement(elem, nextContainerIds);
+				const {containerIds, activeContainerId} = getContainerIds(elem);
+				if (isNavigable(elem, last(containerIds))) {
+					result = focusElement(elem, containerIds, activeContainerId);
 				}
 			}
 
@@ -1359,24 +1375,23 @@ const Spotlight = (function () {
 		 * @public
 		 */
 		move: function (direction, selector) {
-			let elem, containerId;
-
 			direction = direction.toLowerCase();
 			if (!_reverseDirections[direction]) {
 				return false;
 			}
 
-			elem = selector ? parseSelector(selector)[0] : getCurrent();
+			const elem = selector ? parseSelector(selector)[0] : getCurrent();
 			if (!elem) {
 				return false;
 			}
 
-			containerId = getContainerId(elem);
+			const {containerIds, activeContainerId} = getContainerIds(elem);
+			const containerId = last(containerIds);
 			if (!containerId) {
 				return false;
 			}
 
-			return spotNext(direction, elem, containerId);
+			return spotNext(direction, elem, containerId, activeContainerId);
 		},
 
 		/**
@@ -1403,10 +1418,34 @@ const Spotlight = (function () {
 		 * @memberof spotlight.Spotlight.prototype
 		 * @param {String} [containerId] The id of the currently active container. If this is not
 		 *	provided, the root container is set as the currently active container.
+		 * @param {String} [identifier] The id of the container requesting the change. If this is
+		 *	not provided, the provided `containerId` parameter is set as the identifier.
 		 * @public
 		 */
-		setActiveContainer: function (containerId) {
-			_lastContainerId = containerId || spotlightRootContainerName;
+		setActiveContainer: function (containerId, identifier) {
+			if (containerId && !identifier) {
+				identifier = containerId;
+			}
+
+			const lastConfig = _containers.get(_lastContainerId);
+			const lastSpotlightModal = lastConfig && lastConfig.restrict === 'self-only';
+
+			// the current active container is requesting the change, or...
+			if (identifier === _lastContainerId ||
+					// there's a request to change to a specified `containerId` and...
+					(containerId &&
+						// the requested container is not active and...
+						containerId !== _lastContainerId &&
+						// the active container uses a `restrict: 'self-only'` config rule and..
+						lastSpotlightModal &&
+						// the requested container also uses a `restrict: 'self-only'` config rule and..
+						_containers.get(containerId).restrict === 'self-only' &&
+						// the actve container also wraps/contains the requested container
+						(document.querySelector(`[data-container-id='${_lastContainerId}']`) || document).contains(document.querySelector(`[data-container-id='${containerId}']`) || document)) ||
+					// the current active container's config rules will allow focus to move outside its boundaries
+					!lastSpotlightModal) {
+				_lastContainerId = containerId || spotlightRootContainerName;
+			}
 		},
 
 		/**