+++ packages/spotlight/CHANGELOG.md
@@ -6,6 +6,8 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Fixed
 
+- `spotlight/Spotlight` setting the correct active container when calling `focus(containerId)`
+- `spotlight/Spotlight` behavior to follow the various container config rules when navigating between nested containers
 - `spotlight/Spotlight` calling `set()` now properly updates the container config
 - `spotlight/Spotlight` behavior to properly save the last-focused element for nested containers